+++ experimental/packages/opentelemetry-browser-detector/src/BrowserDetector.ts
@@ -19,7 +19,7 @@ import {
   DetectedResource,
   ResourceDetector,
   ResourceDetectionConfig,
-  EMPTY_RESOURCE,
+  emptyResource,
 } from '@opentelemetry/resources';
 import { BROWSER_ATTRIBUTES, UserAgentData } from './types';
 
@@ -30,7 +30,7 @@ class BrowserDetector implements ResourceDetector {
   detect(config?: ResourceDetectionConfig): DetectedResource {
     const isBrowser = typeof navigator !== 'undefined';
     if (!isBrowser) {
-      return EMPTY_RESOURCE;
+      return emptyResource();
     }
     const browserResource: Attributes = getBrowserAttributes();
     return this._getResourceAttributes(browserResource, config);
@@ -53,7 +53,7 @@ class BrowserDetector implements ResourceDetector {
       diag.debug(
         'BrowserDetector failed: Unable to find required browser resources. '
       );
-      return EMPTY_RESOURCE;
+      return emptyResource();
     } else {
       return { attributes: browserResource };
     }