+++ packages/opentelemetry-exporter-jaeger/test/transform.test.ts
@@ -18,7 +18,7 @@ import * as assert from 'assert';
 import { spanToThrift } from '../src/transform';
 import { ReadableSpan } from '@opentelemetry/sdk-trace-base';
 import {
-  EMPTY_RESOURCE,
+  emptyResource,
   resourceFromAttributes,
 } from '@opentelemetry/resources';
 import * as api from '@opentelemetry/api';
@@ -180,7 +180,7 @@ describe('transform', () => {
         links: [],
         events: [],
         duration: [32, 800000000],
-        resource: EMPTY_RESOURCE,
+        resource: emptyResource(),
         instrumentationScope: {
           name: 'default',
           version: '0.0.1',
@@ -254,7 +254,7 @@ describe('transform', () => {
         ],
         events: [],
         duration: [32, 800000000],
-        resource: EMPTY_RESOURCE,
+        resource: emptyResource(),
         instrumentationScope: {
           name: 'default',
           version: '0.0.1',
@@ -302,7 +302,7 @@ describe('transform', () => {
         links: [],
         events: [],
         duration: [32, 800000000],
-        resource: EMPTY_RESOURCE,
+        resource: emptyResource(),
         instrumentationScope: {
           name: 'default',
           version: '0.0.1',