+++ packages/opentelemetry-exporter-zipkin/test/node/zipkin.test.ts
@@ -24,7 +24,7 @@ import {
 } from '@opentelemetry/core';
 import * as api from '@opentelemetry/api';
 import {
-  EMPTY_RESOURCE,
+  emptyResource,
   resourceFromAttributes,
 } from '@opentelemetry/resources';
 import { ZipkinExporter } from '../../src';
@@ -57,7 +57,7 @@ function getReadableSpan() {
     attributes: {},
     links: [],
     events: [],
-    resource: EMPTY_RESOURCE,
+    resource: emptyResource(),
     instrumentationScope: { name: 'default', version: '0.0.1' },
     droppedAttributesCount: 0,
     droppedEventsCount: 0,
@@ -172,7 +172,7 @@ describe('Zipkin Exporter - node', () => {
             attributes: { key3: 'value3' },
           },
         ],
-        resource: EMPTY_RESOURCE,
+        resource: emptyResource(),
         instrumentationScope: { name: 'default', version: '0.0.1' },
         droppedAttributesCount: 0,
         droppedEventsCount: 0,
@@ -198,7 +198,7 @@ describe('Zipkin Exporter - node', () => {
         attributes: {},
         links: [],
         events: [],
-        resource: EMPTY_RESOURCE,
+        resource: emptyResource(),
         instrumentationScope: { name: 'default', version: '0.0.1' },
         droppedAttributesCount: 0,
         droppedEventsCount: 0,
@@ -495,7 +495,7 @@ describe('Zipkin Exporter - node', () => {
           attributes: { key3: 'value3' },
         },
       ],
-      resource: EMPTY_RESOURCE,
+      resource: emptyResource(),
       instrumentationScope: { name: 'default', version: '0.0.1' },
       droppedAttributesCount: 0,
       droppedEventsCount: 0,
@@ -521,7 +521,7 @@ describe('Zipkin Exporter - node', () => {
       },
       links: [],
       events: [],
-      resource: EMPTY_RESOURCE,
+      resource: emptyResource(),
       instrumentationScope: { name: 'default', version: '0.0.1' },
       droppedAttributesCount: 0,
       droppedEventsCount: 0,